feat(auth): refactor signIn function and integrate it into the login page
