+++ apps/web/@utils/apis/auth.ts
@@ -15,6 +15,18 @@ interface SupabaseSignUpRequest {
   password: string;
 }
 
+export const signIn = async (payload: { email: string; password: string }) => {
+  const res = await fetch("/api/auth/signin", {
+    method: "POST",
+    headers: { "Content-Type": "application/json" },
+    body: JSON.stringify(payload),
+    credentials: "include", // [‼️] 쿠키 포함 필수!
+  });
+
+  if (!res.ok) throw new Error("로그인 실패");
+  return res.json();
+};
+
 export const fetchSignIn = async (
   data: SupabaseSignInRequest
 ): Promise<SupabaseSignInResponse> => {
